appending new hybrid response to address response model containers. This breaks DemoUI
